Move Join Meshes to Mesh Tools
I see no point having two files here if we can just move join meshes into the mesh tools file.
