+++ ui/mmd_options.py
@@ -3,7 +3,7 @@ from ..core.register import register_wrap
 from .panel import AvatarToolKit_PT_AvatarToolkitPanel, CATEGORY_NAME
 from ..functions.translations import t
 from ..functions.mmd_functions import *
-from ..functions.join_meshes import AvatarToolKit_OT_JoinAllMeshes
+from ..functions.mesh_tools import AvatarToolKit_OT_JoinAllMeshes
 from ..functions.combine_materials import AvatarToolKit_OT_CombineMaterials
 from ..functions.additional_tools import AvatarToolKit_OT_ApplyTransforms
